Finished backwards propogation with help and understanding from AI. Took way longer to fully wrap my head around it. Even now, I feel like I only understood 40% ish from why the way we do the bakcpropogration

+++ include/activation.hpp
@@ -18,6 +18,6 @@ double mse(const Vec& output, const Vec& target);
 
 double Xavier(int& in);
 
-double HeAdjusted(int& in, double alpha = 0.01);
+double HeAdjusted(double& in, double alpha = 0.01);
 
 #endif
